WIP compute_contact not tested
