+++ docs/grid_model/000_gridmodel_create_model.py
@@ -4,6 +4,7 @@ from compas import json_load
 from compas.datastructures import Mesh
 from compas.geometry import Line
 from compas_model.models import GridModel
+from compas_viewer import Viewer
 
 # =============================================================================
 # JSON file with the geometry of the model.
@@ -20,15 +21,8 @@ model: GridModel = GridModel.from_lines_and_surfaces(columns_and_beams=lines, fl
 # =============================================================================
 # Vizualize
 # =============================================================================
-try:
-    from compas_viewer import Viewer
-
-    viewer = Viewer()
-    viewer.scene.add(model.points)
-    viewer.scene.add(model.lines)
-    viewer.scene.add(model.polygons)
-    viewer.show()
-except ImportError:
-    print("The compas_viewer package is not installed.")
-    print("Please install it by running 'pip install compas_viewer'.")
-    print("Then, try to run this script again.")
+viewer = Viewer()
+viewer.scene.add(model.points)
+viewer.scene.add(model.lines)
+viewer.scene.add(model.polygons)
+viewer.show()
