+++ docs/grid_model/003_gridmodel_add_interfaces_pair_columnhead_column.py
@@ -9,6 +9,7 @@ from compas_model.elements import ScrewElement
 from compas_model.interactions import BooleanModifier
 from compas_model.interactions import SlicerModifier
 from compas_model.models import GridModel
+from compas_viewer import Viewer
 
 # =============================================================================
 # JSON file with the geometry of the model. Datasets: frame.json, crea_4x4.json
@@ -40,13 +41,6 @@ model.add_contact(column_head, column_square)
 # =============================================================================
 # Vizualize
 # =============================================================================
-try:
-    from compas_viewer import Viewer
-
-    viewer = Viewer()
-    viewer.scene.add(model.geometry)
-    viewer.show()
-except ImportError:
-    print("The compas_viewer package is not installed.")
-    print("Please install it by running 'pip install compas_viewer'.")
-    print("Then, try to run this script again.")
+viewer = Viewer()
+viewer.scene.add(model.geometry)
+viewer.show()
