+++ src/compas_model/elements/__init__.py
@@ -1,32 +1,46 @@
-from .element import Element
 from .element import Feature
+from .element import Element
 from .element import reset_computed
-from .block import BlockElement
 from .block import BlockFeature
+from .block import BlockElement
 from .block import BlockGeometry
-from .plate import PlateElement
 from .plate import PlateFeature
-from .column_head_cross import ColumnHeadCrossElement
-from .beam_i_profile import BeamIProfileElement
-from .beam_square import BeamSquareElement
-from .column_round import ColumnRoundElement
-from .column_square import ColumnSquareElement
-from .screw import ScrewElement
+from .plate import PlateElement
+from .column_head import ColumnHeadElement
+from .column_head import ColumnHeadCrossElement
+from .beam import BeamFeature
+from .beam import BeamElement
+from .beam import BeamIProfileElement
+from .beam import BeamSquareElement
+from .column import ColumnFeature
+from .column import ColumnElement
+from .column import ColumnRoundElement
+from .column import ColumnSquareElement
+from .fasteners import FastenersFeature
+from .fasteners import FastenersElement
+from .fasteners import ScrewElement
 
 
 __all__ = [
-    "Element",
-    "Feature",
-    "reset_computed",
-    "BlockElement",
-    "BlockFeature",
-    "BlockGeometry",
-    "PlateElement",
-    "PlateFeature",
-    "ColumnHeadCrossElement",
-    "BeamIProfileElement",
-    "BeamSquareElement",
-    "ColumnRoundElement",
-    "ColumnSquareElement",
-    "ScrewElement",
+    Feature,
+    Element,
+    reset_computed,
+    BlockFeature,
+    BlockElement,
+    BlockGeometry,
+    PlateFeature,
+    PlateElement,
+    ColumnHeadElement,
+    ColumnHeadCrossElement,
+    BeamFeature,
+    BeamElement,
+    BeamIProfileElement,
+    BeamSquareElement,
+    ColumnFeature,
+    ColumnElement,
+    ColumnRoundElement,
+    ColumnSquareElement,
+    FastenersFeature,
+    FastenersElement,
+    ScrewElement,
 ]
