+++ src/compas_model/elements/element.py
@@ -12,7 +12,7 @@ from compas.geometry import Brep
 from compas.geometry import Frame
 from compas.geometry import Shape
 from compas.geometry import Transformation
-from compas_model.interactions import Interaction
+from compas_model.interactions import Interaction, ContactInterface
 from compas_model.materials import Material
 
 if TYPE_CHECKING:
@@ -315,13 +315,15 @@ class Element(Data):
 
         for neighbor in graph.neighbors_in(self.graphnode):
             for interaction in graph.edge_interactions((neighbor, self.graphnode)):  # graph.edge_modififers()
-                # Interaction types: Modifier, Collision, Interface
-                if hasattr(interaction, "modify"):  # Modifier
+                if isinstance(interaction, ContactInterface):
                     modelgeometry = interaction.modify(modelgeometry, elements[neighbor].modelgeometry)
-                elif hasattr(interaction, "collide"):  # Collision:
-                    modelgeometry = interaction.collide(modelgeometry, elements[neighbor].modelgeometry)
-                elif hasattr(interaction, "overlap"):  # Interface:
-                    modelgeometry = interaction.touch(modelgeometry, elements[neighbor].modelgeometry)
+                # # Interaction types: Modifier, Collision, Interface
+                # if hasattr(interaction, "modify"):  # Modifier
+                #     modelgeometry = interaction.modify(modelgeometry, elements[neighbor].modelgeometry)
+                # elif hasattr(interaction, "collide"):  # Collision:
+                #     modelgeometry = interaction.collide(modelgeometry, elements[neighbor].modelgeometry)
+                # elif hasattr(interaction, "overlap"):  # Interface:
+                #     modelgeometry = interaction.touch(modelgeometry, elements[neighbor].modelgeometry)
 
         self.is_dirty = False
